requirements.txt
absl-py==1.2.0
accelerate==0.15.0
black==22.8.0
blis==0.9.1
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1648854167867/work
catalogue==2.0.8
certifi==2022.9.24
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
confection==0.0.1
contextlib2==21.6.0
contourpy==1.0.5
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1652967019182/work
cycler==0.11.0
cymem==2.0.6
diffusers==0.10.2
docker-pycreds==0.4.0
filelock==3.8.0
fonttools==4.37.3
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.11.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.22.0
importlib-metadata==4.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.6.0
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
ml-collections==0.1.1
murmurhash==1.0.8
mypy-extensions==0.4.3
networkx==2.8.6
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
packaging==21.3
pandas==1.5.0
pathspec==0.10.1
pathtools==0.1.2
pathy==0.6.2
Pillow==9.2.0
platformdirs==2.5.2
preshed==3.0.7
promise==2.3
protobuf==3.20.2
psutil==5.9.2
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.9.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
regex==2022.9.13
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
scikit-image==0.19.3
scipy==1.9.1
sentry-sdk==1.9.9
setproctitle==1.3.2
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
smmap==5.0.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
thinc==8.1.1
tifffile==2022.8.12
tokenizers==0.12.1
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.22.1
typer==0.4.2
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
wandb==0.13.3
wasabi==0.10.1
wcwidth==0.2.5
zipp==3.8.1
